first pass of a turret subsystem.
